split out the function to get the exposure URL, simplified the ExposedHelpButton component
