--- src/components/InfoShareItem.tsx
@@ -1,37 +1,0 @@
-import React from 'react';
-import {TouchableOpacity, TouchableOpacityProps} from 'react-native';
-import {Box, Text, Icon, IconProps} from 'components';
-
-interface InfoShareItemProps extends TouchableOpacityProps {
-  onPress: () => void;
-  text: string;
-  icon: IconProps['name'];
-  lastItem?: boolean;
-}
-export const InfoShareItem = ({onPress, text, icon, lastItem, ...touchableProps}: InfoShareItemProps) => (
-  <>
-    <TouchableOpacity activeOpacity={0.6} onPress={onPress} accessibilityRole="button" {...touchableProps}>
-      <Box
-        paddingVertical="s"
-        marginHorizontal="-m"
-        paddingHorizontal="m"
-        flexDirection="row"
-        alignContent="center"
-        justifyContent="space-between"
-        backgroundColor="infoBlockNeutralBackground"
-        borderRadius={5}
-      >
-        <Box flex={1}>
-          <Text variant="bodyText" marginVertical="s" color="overlayBodyText">
-            {text}
-          </Text>
-        </Box>
-
-        <Box alignSelf="center">
-          <Icon size={25} name={icon} />
-        </Box>
-      </Box>
-    </TouchableOpacity>
-    {!lastItem && <Box height={5} marginHorizontal="-m" backgroundColor="overlayBackground" />}
-  </>
-);
