+++ src/components/index.ts
@@ -11,7 +11,6 @@ export * from './CodeInput';
 export * from './Header';
 export * from './Icon';
 export * from './InfoBlock';
-export * from './InfoShareItem';
 export * from './ProgressCircles';
 export * from './Ripple';
 export * from './Text';
