+++ src/screens/home/views/InfoShareView.tsx
@@ -1,13 +1,52 @@
 import React, {useCallback} from 'react';
-import {Linking} from 'react-native';
-import {Box, Text, InfoShareItem} from 'components';
+import {Linking, TouchableOpacity, TouchableOpacityProps} from 'react-native';
+import {Box, Text, Icon, IconProps} from 'components';
 import {useNavigation} from '@react-navigation/native';
-import {useI18n} from 'locale';
+import {useI18n, useRegionalI18n} from 'locale';
 import {captureException} from 'shared/log';
-import {ExposedHelpButton} from 'components/ExposedHelpButton';
+import {getExposedHelpURL} from 'components/ExposedHelpButton';
+import {useStorage} from 'services/StorageService';
+import {isRegionActive} from 'shared/RegionLogic';
+
+interface InfoShareItemProps extends TouchableOpacityProps {
+  onPress: () => void;
+  text: string;
+  icon: IconProps['name'];
+  lastItem?: boolean;
+}
+export const InfoShareItem = ({onPress, text, icon, lastItem, ...touchableProps}: InfoShareItemProps) => (
+  <>
+    <TouchableOpacity activeOpacity={0.6} onPress={onPress} accessibilityRole="button" {...touchableProps}>
+      <Box
+        paddingVertical="s"
+        marginHorizontal="-m"
+        paddingHorizontal="m"
+        flexDirection="row"
+        alignContent="center"
+        justifyContent="space-between"
+        backgroundColor="infoBlockNeutralBackground"
+        borderRadius={5}
+      >
+        <Box flex={1}>
+          <Text variant="bodyText" marginVertical="s" color="overlayBodyText">
+            {text}
+          </Text>
+        </Box>
+
+        <Box alignSelf="center">
+          <Icon size={25} name={icon} />
+        </Box>
+      </Box>
+    </TouchableOpacity>
+    {!lastItem && <Box height={5} marginHorizontal="-m" backgroundColor="overlayBackground" />}
+  </>
+);
 
 export const InfoShareView = () => {
   const i18n = useI18n();
+  const {region} = useStorage();
+  const regionalI18n = useRegionalI18n();
+  const regionActive = isRegionActive(region, regionalI18n.activeRegions);
   const navigation = useNavigation();
   const onPrivacy = useCallback(() => navigation.navigate('Privacy'), [navigation]);
   const onGetCode = useCallback(() => navigation.navigate('NoCode'), [navigation]);
@@ -17,6 +56,11 @@ export const InfoShareView = () => {
   const onHelp = useCallback(() => {
     Linking.openURL(i18n.translate('Info.HelpUrl')).catch(error => captureException('An error occurred', error));
   }, [i18n]);
+  const onExposedHelp = useCallback(() => {
+    Linking.openURL(getExposedHelpURL(region, regionActive, regionalI18n)).catch(error =>
+      captureException('An error occurred', error),
+    );
+  }, [region, regionActive, regionalI18n]);
 
   return (
     <>
@@ -27,7 +71,13 @@ export const InfoShareView = () => {
           text={i18n.translate('Info.GetCode')}
           icon="icon-chevron"
         />
-        <ExposedHelpButton inMenu />
+        <InfoShareItem
+          onPress={onExposedHelp}
+          text={i18n.translate('Home.ExposedHelpCTA')}
+          icon="icon-external-arrow"
+          accessibilityRole="link"
+          accessibilityHint={`${i18n.translate('Home.ExposedHelpCTA')} . ${i18n.translate('Home.ExternalLinkHint')}`}
+        />
         <InfoShareItem
           onPress={onHelp}
           text={i18n.translate('Info.Help')}
